Add Fact.get_serialized_string
This commit adds the ``get_serialized_string`` methid to ``Fact``.
This new method provides a canonical string representation of a
``Fact`` instance. Contrary to ``__str__`` this is about including
*all* relevant data. Contrary to ``__repr__`` we return a 'unicode
string'.

The method take one optional parameter indicating if to include start/end
information or not.

Please note that in order to have a deterministic representation of tags
(which are unordered) we order them by their name before adding them to
the result string.

Closes: LIB-216